+++ boot/i386/libsaio/nbp.c
@@ -3,21 +3,22 @@
  *
  * @APPLE_LICENSE_HEADER_START@
  * 
- * Portions Copyright (c) 1999 Apple Computer, Inc.  All Rights
- * Reserved.  This file contains Original Code and/or Modifications of
- * Original Code as defined in and that are subject to the Apple Public
- * Source License Version 1.1 (the "License").  You may not use this file
- * except in compliance with the License.  Please obtain a copy of the
- * License at http://www.apple.com/publicsource and read it before using
- * this file.
+ * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
+ * 
+ * This file contains Original Code and/or Modifications of Original Code
+ * as defined in and that are subject to the Apple Public Source License
+ * Version 2.0 (the 'License'). You may not use this file except in
+ * compliance with the License. Please obtain a copy of the License at
+ * http://www.opensource.apple.com/apsl/ and read it before using this
+ * file.
  * 
  * The Original Code and all software distributed under the License are
- * distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY KIND, EITHER
+ * distributed on an 'AS IS' basis, WITHOUT WARRANTY OF ANY KIND, EITHER
  * EXPRESS OR IMPLIED, AND APPLE HEREBY DISCLAIMS ALL SUCH WARRANTIES,
  * INCLUDING WITHOUT LIMITATION, ANY WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE OR NON- INFRINGEMENT.  Please see the
- * License for the specific language governing rights and limitations
- * under the License.
+ * FITNESS FOR A PARTICULAR PURPOSE, QUIET ENJOYMENT OR NON-INFRINGEMENT.
+ * Please see the License for the specific language governing rights and
+ * limitations under the License.
  * 
  * @APPLE_LICENSE_HEADER_END@
  */
@@ -69,11 +70,11 @@ static long NBPLoadFile(CICell ih, char * filePath)
 	cmd.bufferSize = TFTP_LEN;
 	cmd.buffer     = TFTP_ADDR;
 
-    verbose("Loading file: %s\n", filePath);
+	verbose("Loading file: %s\n", filePath);
 
 	ret = nbp(nbpCommandTFTPReadFile, (nbpCommand_u *) &cmd);
 
-    return (ret == nbpStatusSuccess) ? cmd.bufferSize : -1;
+    return (ret == nbpStatusSuccess) ? (long)cmd.bufferSize : -1;
 }
 
 /*==========================================================================
@@ -113,6 +114,5 @@ BVRef nbpScanBootVolumes( int biosdev, int * countPtr )
             gNetBVR->fs_getdirentry = NBPGetDirEntry;
         }
     }
-
     return gNetBVR;
 }
