+++ developer_cmds/indent/indent.c
@@ -1150,7 +1150,7 @@ check_type:
 					while ((c = getc(input)) == '\n');
 					ungetc(c, input);
 				}
-				if (ifdef_level < sizeof state_stack / sizeof state_stack[0]) {
+				if (ifdef_level < (int)(sizeof state_stack / sizeof state_stack[0])) {
 					match_state[ifdef_level].tos = -1;
 					state_stack[ifdef_level++] = ps;
 				} else
